requirements.txt
Pillow==10.3.0
streamlit==1.34.0
~treamlit==1.12.0
